Main

+++ TrabajoPracticoFinal/src/interfaces/IDescuento.java
@@ -2,8 +2,8 @@ package interfaces;
 
 public interface IDescuento {
 
-	public static final int DESC_TARJETA = 0;
-	public static final int DESC_EFECTIVO = 20;
-	public static final int DESC_MERCADOPAGO = 5;
-	public static final int DESC_CUENTADNI = 15;
+	public static final double DESC_TARJETA = 0;
+	public static final double DESC_EFECTIVO = 20;
+	public static final double DESC_MERCADOPAGO = 5;
+	public static final double DESC_CUENTADNI = 15;
 }
